+++ web/src/components/Chip/Chip.jsx
@@ -4,10 +4,14 @@ import PropTypes from "prop-types";
 import "./Chip.scss";
 import StopProp from "@/components/StopProp/StopProp";
 
-const Chip = ({ children, onClick, style }) => {
+const Chip = ({ children, onClick, style, shake}) => {
+    let className = "chip"
+    if(shake){
+        className += " shake"
+    }
     return (
         <StopProp>
-            <div className="chip" onClick={onClick} style={style}>
+            <div className={className} onClick={onClick} style={style}>
                 {children}
             </div>
         </StopProp>
@@ -17,6 +21,7 @@ Chip.propTypes = {
     children: PropTypes.any.isRequired,
     onClick: PropTypes.func,
     style: PropTypes.object,
+    pulse: PropTypes.bool,
 };
 
 export default Chip;